monitor/requirements/prod.txt
celery
celery[redis]
kubernetes
pyyaml
